# openapi.json
{"openapi":"3.1.0","info":{"title":"Kajima App API","version":"0.1.0"},"servers":[{"url":"/dev"}],"paths":{"/auth/oauth/{provider_id}":{"get":{"tags":["auth"],"summary":"Oauth Login","description":"Redirect user to the OAuth provider login page.","operationId":"oauth_login_auth_oauth__provider_id__get","parameters":[{"name":"provider_id","in":"path","required":true,"schema":{"type":"string","title":"Provider Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/oauth/{provider_id}/callback":{"get":{"tags":["auth"],"summary":"Oauth Callback","description":"Handle the OAuth provider callback.","operationId":"oauth_callback_auth_oauth__provider_id__callback_get","parameters":[{"name":"provider_id","in":"path","required":true,"schema":{"type":"string","title":"Provider Id"}},{"name":"error","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},{"name":"code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code"}},{"name":"state","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"get":{"tags":["auth"],"summary":"Login","description":"Handle the login request.","operationId":"login_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/get_access_token":{"get":{"tags":["auth"],"summary":"Get Access Token","description":"Get the access token from the request.","operationId":"get_access_token_get_access_token_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearerWithCookie":[]}]}},"/logout":{"get":{"tags":["auth"],"summary":"Logout","operationId":"logout_logout_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health check for GET method","description":"This is a health check API for GET Method. This API will always return message I'm active now","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{},"example":{"message":"I'm active now"}}}}}}},"/return_same":{"post":{"summary":"Health check for POST method","description":"This is a health check API for POST Method. This API will return the same message as request.","operationId":"return_same_return_same_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Say"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{},"example":{"message":"Hello"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trigger_classification":{"post":{"summary":"Trigger classification for image included in record id","description":"This API will trigger the classification for image included in record id. This API will return the same message as request.","operationId":"trigger_classification_trigger_classification_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_trigger_classification_trigger_classification_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{},"example":{"user_id":"User ID","record_id":"Record ID","message":"Classification triggered successfully"}}}},"404":{"description":"Not Found","content":{"application/json":{"example":{"message":"Record not found"}}}},"403":{"description":"Forbidden","content":{"application/json":{"example":{"message":"Permission denied. Please check your Kintone API token."}}}},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"message":"Unauthorized. Please check your access token."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearerWithCookie":[]}]}}},"components":{"schemas":{"Body_trigger_classification_trigger_classification_post":{"properties":{"user_id":{"type":"string","title":"User Id"},"record_id":{"type":"string","title":"Record Id"}},"type":"object","required":["user_id","record_id"],"title":"Body_trigger_classification_trigger_classification_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Say":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"Say"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearerWithCookie":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/login"}}}}}}